Rename 'visualizer' to 'viewer' in all locations
