+++ extensions/ql-vscode/package.json
@@ -38,7 +38,7 @@
     "onView:codeQLDatabases",
     "onView:codeQLQueryHistory",
     "onView:codeQLAstViewer",
-    "onView:codeQLEvalLogVisualizer",
+    "onView:codeQLEvalLogViewer",
     "onView:test-explorer",
     "onCommand:codeQL.checkForUpdatesToCLI",
     "onCommand:codeQL.authenticateToGitHub",
@@ -528,15 +528,15 @@
       },
       {
         "command": "codeQLQueryHistory.showEvalLog",
-        "title": "Show Evaluator Log (Raw)"
+        "title": "Show Evaluator Log (Raw JSON)"
       },
       {
         "command": "codeQLQueryHistory.showEvalLogSummary",
-        "title": "Show Evaluator Log (Summary)"
+        "title": "Show Evaluator Log (Summary Text)"
       },
       {
-        "command": "codeQLQueryHistory.showEvalLogVisualizer",
-        "title": "Show Evaluator Log (Visualizer)"
+        "command": "codeQLQueryHistory.showEvalLogViewer",
+        "title": "Show Evaluator Log (UI)"
       },
       {
         "command": "codeQLQueryHistory.cancel",
@@ -615,8 +615,8 @@
         }
       },
       {
-        "command": "codeQLEvalLogVisualizer.clear",
-        "title": "Clear Visualizer",
+        "command": "codeQLEvalLogViewer.clear",
+        "title": "Clear Viewer",
         "icon": {
           "light": "media/light/clear-all.svg",
           "dark": "media/dark/clear-all.svg"
@@ -696,8 +696,8 @@
           "group": "navigation"
         },
         {
-          "command": "codeQLEvalLogVisualizer.clear",
-          "when": "view == codeQLEvalLogVisualizer",
+          "command": "codeQLEvalLogViewer.clear",
+          "when": "view == codeQLEvalLogViewer",
           "group": "navigation"
         }
       ],
@@ -773,7 +773,7 @@
           "when": "codeql.supportsEvalLog && viewItem == rawResultsItem || codeql.supportsEvalLog && viewItem == interpretedResultsItem || codeql.supportsEvalLog && viewItem == cancelledResultsItem"
         },
         {
-          "command": "codeQLQueryHistory.showEvalLogVisualizer",
+          "command": "codeQLQueryHistory.showEvalLogViewer",
           "group": "9_qlCommands",
           "when": "config.codeQL.canary && codeql.supportsEvalLog && viewItem == rawResultsItem || config.codeQL.canary && codeql.supportsEvalLog && viewItem == interpretedResultsItem || config.codeQL.canary && codeql.supportsEvalLog && viewItem == cancelledResultsItem"
         },
@@ -999,7 +999,7 @@
           "when": "false"
         },
         {
-          "command": "codeQLQueryHistory.showEvalLogVisualizer",
+          "command": "codeQLQueryHistory.showEvalLogViewer",
           "when": "false"
         },
         {
@@ -1071,7 +1071,7 @@
           "when": "false"
         },
         {
-          "command": "codeQLEvalLogVisualizer.clear",
+          "command": "codeQLEvalLogViewer.clear",
           "when": "false"
         },
         {
@@ -1142,9 +1142,8 @@
           "name": "AST Viewer"
         },
         {
-          "id": "codeQLEvalLogVisualizer",
-          "name": "Evaluator Log Visualizer",
-          "visibility": "collapsed",
+          "id": "codeQLEvalLogViewer",
+          "name": "Evaluator Log Viewer",
           "when": "config.codeQL.canary"
         }
       ]
@@ -1163,8 +1162,8 @@
         "contents": "Add a CodeQL database:\n[From a folder](command:codeQLDatabases.chooseDatabaseFolder)\n[From an archive](command:codeQLDatabases.chooseDatabaseArchive)\n[From a URL (as a zip file)](command:codeQLDatabases.chooseDatabaseInternet)\n[From LGTM](command:codeQLDatabases.chooseDatabaseLgtm)"
       },
       {
-        "view": "codeQLEvalLogVisualizer",
-        "contents": "Run the 'Show Evaluator Log (Visualizer)' command on a CodeQL query run in the Query History view."
+        "view": "codeQLEvalLogViewer",
+        "contents": "Run the 'Show Evaluator Log (UI)' command on a CodeQL query run in the Query History view."
       }
     ]
   },
