+++ extensions/ql-vscode/src/query-results.ts
@@ -218,7 +218,7 @@ export class LocalQueryInfo {
   public completedQuery: CompletedQueryInfo | undefined;
   public evalLogLocation: string | undefined;
   public evalLogSummaryLocation: string | undefined;
-  public evalLogVisualizerData: EvalLogData[] | undefined;
+  public evalLogViewerData: EvalLogData[] | undefined;
 
   /**
    * Note that in the {@link slurpQueryHistory} method, we create a FullQueryInfo instance
