+++ extensions/ql-vscode/src/vscode-tests/no-workspace/remote-queries/remote-query-history.test.ts
@@ -19,7 +19,7 @@ import { getErrorMessage } from '../../../pure/helpers-pure';
 import { HistoryItemLabelProvider } from '../../../history-item-label-provider';
 import { RemoteQueriesManager } from '../../../remote-queries/remote-queries-manager';
 import { InterfaceManager } from '../../../interface';
-import { EvalLogVisualizer } from '../../../eval-log-visualizer';
+import { EvalLogViewer } from '../../../eval-log-viewer';
 
 /**
  * Tests for remote queries and how they interact with the query history manager.
@@ -94,7 +94,7 @@ describe('Remote queries and query history manager', function() {
       {} as DatabaseManager,
       localQueriesInterfaceManagerStub,
       remoteQueriesManagerStub,
-      {} as EvalLogVisualizer,
+      {} as EvalLogViewer,
       STORAGE_DIR,
       {
         globalStorageUri: Uri.file(STORAGE_DIR),
